working commandline interface and simple options CSV does not work but tries to
