+++ crobat/LOB_funcs.py
@@ -1,63 +1,45 @@
 import pandas as pd
 import copy
-from bisect import bisect
+import bisect
 import numpy as np
-#import fixed_tick_LOB as ftLOB
-# LOB_funcs.py contains the functions that are used in the class that interacts with the WebSocket. 
-# The initial functions price
-
-def price_match(x,y):
-    if y == x:
-        return True
-    else:
-        return False            
-#This function converts my ...
-def convert_array_to_list_dict(history, pos_limit=5):
-    temp_dict = {}
-    return_list = []
-    for i in range(len(history)):
-        temp_dict.update({"time":history[i][0]})
-        for n in range(pos_limit):
-            temp_dict.update({str(n):history[i][1][n][1]})
-        return_list.append(temp_dict)
-        temp_dict = {}
-    return return_list
-
-def pd_excel_save(title, hist_obj_dict):
-    hist_obj_df = pd.DataFrame(hist_obj_dict)
-    hist_obj_df = hist_obj_df[1:]
-    writer = pd.ExcelWriter(title, engine='xlsxwriter')
-    hist_obj_df.to_excel(writer, sheet_name='Sheet1')
-    writer.save()
+import history_funcs as hf
 
 class history(object):
-    # init creates the object that I will start modifying 
     def __init__(self):
-        self.bid_history=[]#snapshot.x_history #an empty list that will hold the history of the list tuples we're going to be working on 
+        self.bid_history=[]
         self.ask_history=[]
-        self.snapshot_bid = []# snapshot.x #an empty list that will hold the current instance of the list tuples we're going to be working on 
+        self.signed_history = []
+        self.snapshot_bid = []
         self.snapshot_ask = []
-        self.bid_events = []# list two elemtnts classification, ,price, and size ['insertion', 0.3105 XRP-USD, 12345 XRP] 
+        self.snapshot_signed = []
+        self.bid_events = []
         self.ask_events = []
+        self.signed_events = []
         self.order_type = None
         self.token = False
         self.position = 0
         self.event_size = 0
+        #self.bid_price_hist = []
+        #self.ask_price_hist = []
+        #self.signed_price_hist = []
 
     def initialize_snap_events(self, msg, time):
         time = time
-        self.snapshot_bid = msg['bids'][:3800]#[:3800] #creates the list of [[price, size], ... [price, size]] which is what we'll be working on
+        self.snapshot_bid = msg['bids'][:3800]
         self.snapshot_ask = msg['asks'][:3800] 
         self.bid_range = [float(self.snapshot_bid[i][0]) for i in range(len(self.snapshot_bid))]
-        self.ask_range = [float(self.snapshot_ask[i][0]) for i in range(len(self.snapshot_ask))] # x_range is the range in prices from our initial observation
-        self.bid_volm  = [float(self.snapshot_bid[i][1]) for i in range(len(self.snapshot_bid))] # x_volm is the associated volum from our initial observation 
-        self.ask_volm  = [float(self.snapshot_ask[i][1]) for i in range(len(self.snapshot_ask))]
+        self.ask_range = [float(self.snapshot_ask[i][0]) for i in range(len(self.snapshot_ask))] 
+        self.min_dec = hf.get_min_dec(0.01,self.bid_range[0])
+        self.bid_volm  = np.round([float(self.snapshot_bid[i][1]) for i in range(len(self.snapshot_bid))],decimals = self.min_dec)
+        self.ask_volm  = np.round([float(self.snapshot_ask[i][1]) for i in range(len(self.snapshot_ask))],decimals = self.min_dec)
         self.snapshot_bid = [[self.bid_range[i], self.bid_volm[i]] for i in range(len(self.snapshot_bid))]
-        self.snapshot_ask = [[self.ask_range[i], self.ask_volm[i]] for i in range(len(self.snapshot_ask))] #convert x from str into floats #convert x from str into floats
-        self.bid_history.append([time, self.snapshot_bid]) # append this initial instance of x to history
-        self.ask_history.append([time, self.snapshot_ask]) # append this initial instance of x to history
-
-    # event that will happen when a market order arrives defined in the ticker class, on message function
+        self.snapshot_ask = [[self.ask_range[i], self.ask_volm[i]] for i in range(len(self.snapshot_ask))] 
+        self.snapshot_signed = [[i[0], -1*i[1]] for i in self.snapshot_bid][::-1] + self.snapshot_ask
+        self.bid_history.append([time, self.snapshot_bid]) 
+        self.ask_history.append([time, self.snapshot_ask]) 
+        self.signed_history.append([time, self.snapshot_signed])
+        #print(self.snapshot_bid)
+        #self.round_digits = 0.01/self.ask_range[-1] # smallest rize i'll allow
     def add_market_order_message(self, message, events):
         index_start = 0
         index_stop = len(events) 
@@ -77,13 +59,13 @@ class history(object):
     # functions that modify the objects in __init__ on the arrival of L2 update. 
     def remove_price_level(self, snap_array, level_depth, match_index):
         if level_depth == 0 and match_index:            
-            snap_array = [[snap_array[i][0], snap_array[i][1]] for i in range(len(snap_array)) if snap_array[i][1] != 0]
+            del snap_array[match_index[0]]
             self.token=True
         return snap_array
 
     def update_level_depth(self, snap_array, level_depth, match_index, pre_level_depth):
         if match_index: 
-            update_index = match_index[0] # position of where the change happened
+            update_index = match_index[0] 
             self.event_size = abs(snap_array[update_index][1] - level_depth)
             pre_level_depth = snap_array[update_index][1]
             if snap_array[update_index][1] < level_depth:
@@ -92,7 +74,7 @@ class history(object):
                 self.order_type = 'cancelation'
             else:
                 pass
-            snap_array[update_index][1] = level_depth #sets the updated volm
+            snap_array[update_index][1] = level_depth 
             self.token=True
             self.position = update_index
         else:
@@ -104,24 +86,25 @@ class history(object):
             self.order_type = 'cancelation' if level_depth == 0 else self.order_type
             pre_level_depth = 0
             self.event_size = level_depth
-            if price_level > max(self.bid_range): #it the message brought back a price level that was not in the original self.x we update it in our self.x_range, and self.x lists
-                self.bid_range.append(price_level) # update the x_range
-                self.snapshot_bid.append([price_level,level_depth]) # add the new pricelevel, and volm pair to self.x
+            if price_level > self.bid_range[0]: 
+                self.bid_range.insert(0,price_level) 
+                self.snapshot_bid.insert(0,[price_level,level_depth]) 
                 self.order_type = "insertion"
                 self.token = True
                 self.position = 0
-            elif price_level < min(self.bid_range): #when it's below the range, we are not interested in it (orderbook stuff dwai)
+            elif price_level < self.bid_range[-1]: 
                 self.token=False
-                self.position = None
-            else: # sometimes I get a new price that is between the min and max of x_range, but not a member of the list, this says that its okay to add.             
-                self.bid_range.append(price_level) #adds the new price to x_range
-                sorted(self.bid_range)[::-1]
-                sorted(self.snapshot_bid)[::-1]
-                self.bid_range= list(set(self.bid_range))
-                self.position = bisect(self.bid_range, price_level)
-                self.snapshot_bid[self.position:self.position] = [[price_level, level_depth]]
-                self.order_type = "insertion"
-                self.token=True
+                self.position = len(self.bid_range)
+            else: 
+                self.position = bisect.bisect(self.bid_range, price_level)
+                if self.position == 0:
+                    self.token=False
+                    print("encountered problem ON BUY SIDE assigning correct bisect point for price level", price_level, "at position", self.position)
+                else:
+                    self.snapshot_bid.insert(self.position, [price_level, level_depth])
+                    self.bid_range.insert(self.position, price_level)
+                    self.order_type = "insertion"
+                    self.token=True
         return self.snapshot_bid, self.bid_range, pre_level_depth
 
     def update_price_index_sell(self, level_depth, price_level, pre_level_depth):
@@ -129,72 +112,176 @@ class history(object):
             self.order_type = 'cancelation' if level_depth == 0 else self.order_type
             pre_level_depth = 0
             self.event_size = level_depth
-            if price_level < min(self.ask_range): #it the message brought back a price level that was not in the original self.x we update it in our self.x_range, and self.x lists
-                self.ask_range.append(price_level) # update the x_range
-                self.snapshot_ask.append([price_level,level_depth]) # add the new pricelevel, and volm pair to self.x
+            if price_level < min(self.ask_range): 
+                self.ask_range.insert(0, price_level) 
+                self.snapshot_ask.insert(0, [price_level,level_depth]) 
                 self.order_type = "insertion"
                 self.token = True
                 self.position = 0
-            elif price_level > max(self.ask_range): #when it's below the range, we are not interested in it (orderbook stuff dwai)
+            elif price_level > max(self.ask_range): 
                 self.token=False
-                self.position = None
-            else: #check to see that I have not alread made a change            
-                self.ask_range.append(price_level) #adds the new price to x_range
-                sorted(self.ask_range)
-                sorted(self.snapshot_ask)
-                self.ask_range = list(set(self.ask_range))
-                self.position = bisect(self.ask_range, price_level)
-                self.snapshot_ask[self.position:self.position] = [[price_level, level_depth]]
-                self.order_type = "insertion"
-                self.token=True
-                #print("the ask updated position is", position)
+                self.position = len(self.ask_range)
+            else:
+                self.position = bisect.bisect(self.ask_range, price_level) 
+                if self.position == 0:
+                    self.token=False
+                    print("encountered problem ON SELL SIDE with assigning correct bisect point for price level", price_level, "at position", self.position)
+                else: 
+                    self.snapshot_ask.insert(self.position, [price_level, level_depth])
+                    self.ask_range.insert(self.position, price_level)
+                    self.order_type = "insertion"
+                    self.token=True
         return self.snapshot_ask, self.ask_range, pre_level_depth
 
     def update_snapshot_bid(self):
-        sorted(self.snapshot_bid)#.sort(key=lambda x: x[0])[::-1]
-        self.snapshot_bid[::-1]# = snapshot_array[::-1]
         self.bid_range = [self.snapshot_bid[i][0] for i in range(len(self.snapshot_bid))]
         return self.snapshot_bid, self.bid_range
 
     def update_snapshot_ask(self):
-        sorted(self.snapshot_ask)#.sort(key=lambda x: x[0])[::-1]
         self.ask_range= [self.snapshot_ask[i][0] for i in range(len(self.snapshot_ask)) ]
         return self.snapshot_ask, self.ask_range
 
     def trim_coordinator(self, position, bound):
-        try:
-            if position>bound:
-                self.token = False
-            else:
-                pass
-        except:
+        if position>bound:
             self.token = False
+        else:
+            pass
         return self.token
 
     def append_snapshot_bid(self, time, price_level):
-        mid_price = 0.5*(max(self.bid_range) + min(self.ask_range))
-        temp_snap = copy.deepcopy(self.snapshot_bid[:5])
-        spread = min(self.ask_range) - max(self.bid_range)
+        mid_price = 0.5*(self.bid_range[0] + self.ask_range[0])
+        temp_snap = copy.deepcopy(self.snapshot_bid[:6])
+        spread = self.ask_range[0] - self.bid_range[0]
         self.bid_history.append([time, temp_snap])
-        self.bid_events.append([time, self.order_type, price_level, self.event_size, self.position, mid_price, spread])
+        self.bid_events.append([time, self.order_type, price_level, self.event_size, self.position+1, mid_price, spread])
 
     def append_snapshot_ask(self, time, price_level):
-        mid_price = 0.5*(max(self.bid_range) + min(self.ask_range))
-        spread = min(self.ask_range) - max(self.bid_range)
-        temp_snap = copy.deepcopy(self.snapshot_ask[:5])
+        mid_price = 0.5*(self.bid_range[0] + self.ask_range[0])
+        spread = self.ask_range[0] - self.bid_range[0]
+        temp_snap = copy.deepcopy(self.snapshot_ask[:6])
         self.ask_history.append([time, temp_snap])
-        self.ask_events.append([time, self.order_type, price_level, self.event_size, self.position, mid_price, spread])
+        self.ask_events.append([time, self.order_type, price_level, self.event_size, self.position+1, mid_price, spread])
+    
+    def append_signed_book(self, time, price_level, side):
+        mid_price = 0.5*(self.bid_range[0] + self.ask_range[0])
+        spread = self.ask_range[0] - self.bid_range[0]
+        sign = hf.set_sign(self.event_size, side, self.order_type)
+        self.event_size *= sign
+        self.position = hf.set_signed_position(self.position, side)
+        temp_snap_bid = [[i[0],-1*i[1]] for i in copy.deepcopy(self.snapshot_bid[:6])][::-1]
+        temp_snap_ask = copy.deepcopy(self.snapshot_ask[:6])        
+        self.signed_history.append([time, temp_snap_bid + temp_snap_ask])
+        self.signed_events.append([time, self.order_type, price_level, self.event_size, self.position, side, mid_price, spread])
 
-    def check_mkt_can_overlap(self, events):
+    def check_mkt_can_overlap(self, events, order_type):
         set_of_order_of_arrival = [['market', 'cancelation'],['cancelation', 'market']]
         if len(events)>2:
             last_two = events[-2:]
             orders = [last_two[0][1],last_two[1][1]]
             sizes = [last_two[0][3], last_two[1][3]]
-            if (sizes[0] == sizes[1]) and (orders in set_of_order_of_arrival):
-                i = orders.index('cancelation') -2
-                del events[i]
-    
+            if sizes[0] - sizes[1] == float(0):
+                if order_type == 'market':
+                    #print("1a.market message initiated deleteing", events[-2])
+                    del events[-2]
+                elif (order_type == 'cancelation') and (orders[0] == 'market'):
+                    #print("1b. cancelation message initiated", events[-1])
+                    del events[-1]
+                else:
+                    #print("1c.sizes agree but neither mkt,can or can,mkt received")
+                    pass
+            else:
+                #print("sizes don't agree", sizes)
+                pass
+
+    #### Accessors #####
+    def last_inserted_order(self, side="signed"): #these args side, 
+        out = []
+        if side == "buy":
+            event_list = self.bid_events
+        elif side == "sell":
+            event_list = self.ask_events
+        else:
+            event_list = self.signed_events
+
+        len_check = min(len(event_list),30)
+
+        for i in range(len(event_list[-len_check:])):
+            if event_list[::-1][i][1] == "insertion":
+                out = event_list[::-1][i]
+                break
+            else:
+                out = []
+                print('no LO was found in last 30 messages')
+        return out
+
+
+    def last_canceled_order(self, side="signed"):
+        out = []
+        if side == "buy":
+            event_list = self.bid_events
+        elif side == "sell":
+            event_list = self.ask_events
+        else:
+            event_list = self.signed_events
+
+        len_check = min(len(event_list),30)
+
+        for i in range(len(event_list[-len_check:])):
+            if event_list[::-1][i][1] == "insertion":
+                out = event_list[::-1][i]
+                break
+            else:
+                out = []
+                print('no CO was found in last 30 messages')
+        return out
+
+
+    def last_market_order(self, side="signed"):
+        out = []
+        if side == "buy":
+            event_list = self.bid_events
+        elif side == "sell":
+            event_list = self.ask_events
+        else:
+            event_list = self.signed_events
+
+        len_check = min(len(event_list),30)
+
+        for i in range(len(event_list[-len_check:])):
+            if event_list[::-1][i][1] == "market":
+                out = event_list[::-1][i]
+                break
+            else:
+                out = []
+                print('no MO was found in last 30 messages')
+        return out
+
+    def last_orderbook_image(self, side="signed"):
+        if side == "buy":
+            orderbook_list = self.bid_history
+        elif side == "sell":
+            orderbook_list = self.ask_history
+        else:
+            orderbook_list = self.signed_history
+
+        return orderbook_list[-1]
+
+    def last_market_depth(self, side, pos_range='all'): 
+        if side == "buy":
+            orderbook_snap = self.bid_history[-1][1]
+        elif side =="sell":
+            orderbook_snap = self.ask_history[-1][1]
+        else:
+            print("error, no side selected. please pick bid or ask")
+        if pos_range == 'all':
+            pass
+        else:
+            orderbook_snap = orderbook_snap[:pos_range]
+        depth = 0    
+        for i in range(len(orderbook_snap)):
+            depth += orderbook_snap[i][1] * orderbook_snap[i][0]
+        return depth     
+
 def UpdateSnapshot_bid_Seq(hist_obj, time, side, price_level, level_depth, pre_level_depth, price_match_index):
     hist_obj.snapshot_bid, pre_level_depth = hist_obj.update_level_depth(hist_obj.snapshot_bid, level_depth, price_match_index, pre_level_depth)        
     hist_obj.snapshot_bid = hist_obj.remove_price_level(hist_obj.snapshot_bid, level_depth, price_match_index) # needs price range update, 
@@ -203,7 +290,19 @@ def UpdateSnapshot_bid_Seq(hist_obj, time, side, price_level, level_depth, pre_l
     hist_obj.token = hist_obj.trim_coordinator(hist_obj.position, 5)
     if hist_obj.token:
         hist_obj.append_snapshot_bid(time, price_level)
-        hist_obj.check_mkt_can_overlap(hist_obj.bid_events)
+        hist_obj.append_signed_book(time, price_level, side)
+        hist_obj.check_mkt_can_overlap(hist_obj.bid_events, hist_obj.order_type)
+        hist_obj.check_mkt_can_overlap(hist_obj.signed_events, hist_obj.order_type)
+        if hist_obj.bid_events[-1][-1] > 100:
+            print("it happened on this message", hist_obj.bid_events[-1])
+            print("the type of addition was: ", hist_obj.order_type)
+            print("   ")
+            print("the previous bid event was:", hist_obj.bid_events[-2])
+            print("the previous ask event was:", hist_obj.ask_events[-2])
+            print("the previous bid book looked like", hist_obj.bid_history[-2])
+            print("the previous ask book looked like", hist_obj.ask_history[-2])
+            print("the bid book looks like NOW:", hist_obj.bid_history[-1])
+            print("the ask book looks like NOW:", hist_obj.ask_history[-1])
 
 def UpdateSnapshot_ask_Seq(hist_obj, time, side, price_level, level_depth, pre_level_depth, price_match_index):
     hist_obj.snapshot_ask, pre_level_depth = hist_obj.update_level_depth(hist_obj.snapshot_ask, level_depth, price_match_index, pre_level_depth)        
@@ -213,5 +312,36 @@ def UpdateSnapshot_ask_Seq(hist_obj, time, side, price_level, level_depth, pre_l
     hist_obj.token = hist_obj.trim_coordinator(hist_obj.position, 5)
     if hist_obj.token:
         hist_obj.append_snapshot_ask(time, price_level)
-        hist_obj.check_mkt_can_overlap(hist_obj.ask_events)
-        
+        hist_obj.append_signed_book(time, price_level, side)
+        hist_obj.check_mkt_can_overlap(hist_obj.ask_events, hist_obj.order_type)
+        hist_obj.check_mkt_can_overlap(hist_obj.signed_events, hist_obj.order_type)
+        if hist_obj.ask_events[-1][-1] > 100:
+            print("it happened on this message", hist_obj.ask_events[-1])
+            print("the type of addition was: ", hist_obj.order_type)
+            print("   ")
+            print("the previous bid event was:", hist_obj.bid_events[-2])
+            print("the previous ask event was:", hist_obj.ask_events[-2])
+            print("the previous bid book looked like", hist_obj.bid_history[-2])
+            print("the previous ask book looked like", hist_obj.ask_history[-2])
+            print("the bid book looks like NOW:", hist_obj.bid_history[-1])
+            print("the ask book looks like NOW:", hist_obj.ask_history[-1])
+
+def price_match(x,y):
+    if y == x:
+        return True
+    else:
+        return False 
+
+# def get_min_dec(min_currency_denom, min_asset_value):
+#     min_tradable_amount = min_currency_denom/min_asset_value
+#     min_dec = 0
+#     while min_tradable_amount < 1:
+#         min_tradable_amount*=10
+#         min_dec += 1
+#         if min_dec_out > 15:
+#             break 
+#     return min_dec_out
+
+
+
+
