+++ docs/manual/manual.aux
@@ -1,0 +1,37 @@
+\relax 
+\providecommand\hyper@newdestlabel[2]{}
+\providecommand\HyperFirstAtBeginDocument{\AtBeginDocument}
+\HyperFirstAtBeginDocument{\ifx\hyper@anchor\@undefined
+\global\let\oldcontentsline\contentsline
+\gdef\contentsline#1#2#3#4{\oldcontentsline{#1}{#2}{#3}}
+\global\let\oldnewlabel\newlabel
+\gdef\newlabel#1#2{\newlabelxx{#1}#2}
+\gdef\newlabelxx#1#2#3#4#5#6{\oldnewlabel{#1}{{#2}{#3}}}
+\AtEndDocument{\ifx\hyper@anchor\@undefined
+\let\contentsline\oldcontentsline
+\let\newlabel\oldnewlabel
+\fi}
+\fi}
+\global\let\hyper@last\relax 
+\gdef\HyperFirstAtBeginDocument#1{#1}
+\providecommand\HyField@AuxAddToFields[1]{}
+\providecommand\HyField@AuxAddToCoFields[2]{}
+\bibstyle{plain}
+\@input{sections/Introduction/Introduction.aux}
+\@input{sections/Structure/Order_Books.aux}
+\@input{sections/Context/context.aux}
+\@input{sections/Modules/Modules.aux}
+\@input{sections/Contributing/contributing.aux}
+\@input{sections/Credits/credits.aux}
+\@input{sections/License/License.aux}
+\@input{sections/Appendix/Appendix.aux}
+\@input{sections/Appendix/function_template.aux}
+\bibdata{sections/Bibliography/manual}
+\bibcite{Huang:2020}{{1}{}{{}}{{}}}
+\bibcite{CarteaJaimungalWang2020:SpoofingPriceManipulation}{{2}{}{{}}{{}}}
+\bibcite{Ohara:MMTheory}{{3}{}{{}}{{}}}
+\bibcite{ContKukanovStoikov2012:PriceImpactOBE}{{4}{}{{}}{{}}}
+\bibcite{ContStoikovTalreja2009:StochModelOBD}{{5}{}{{}}{{}}}
+\bibcite{Silyantev2019:Orderflowanalysis}{{6}{}{{}}{{}}}
+\providecommand\NAT@force@numbers{}\NAT@force@numbers
+\gdef \@abspage@last{43}
