+++ docs/manual/manual.bbl
@@ -1,0 +1,34 @@
+\begin{thebibliography}{1}
+
+\bibitem{Huang:2020}
+Weibing Huang, Charles-Albert Lehalle, and Mathieu Rosenbaum.
+\newblock Simulating and analyzing order book data: The queue-reactive model.
+\newblock {\em Journal of the American Statistical Association},
+  110(509):107--122, 2015.
+
+\bibitem{CarteaJaimungalWang2020:SpoofingPriceManipulation}
+Álvaro Cartea, Sebastian Jaimungal, and Yixuan Wang.
+\newblock Spoofing and price manipulation in order-driven markets.
+\newblock {\em Applied Mathematical Finance}, 27(1-2):67--98, 2020.
+
+\bibitem{Ohara:MMTheory}
+Maureen O'Hara.
+\newblock {\em Market Microstructure Theory}.
+\newblock John Wiley \& Sons Inc., 1998.
+
+\bibitem{ContKukanovStoikov2012:PriceImpactOBE}
+Arseniy~Kukanov Rama~Cont and Sasha Stoikov.
+\newblock The price impact of order book events.
+\newblock {\em SSRN}, pages 1--32, 2012.
+
+\bibitem{ContStoikovTalreja2009:StochModelOBD}
+Sasha~Stoikov Rama~Cont and Rishi Talreja.
+\newblock A stochasic model for order book dynamics.
+\newblock {\em SSRN}, pages 1--33, 2009.
+
+\bibitem{Silyantev2019:Orderflowanalysis}
+Edward Silyantev.
+\newblock Order flow analysis of crpytocurrency markets.
+\newblock {\em Digital Finance}, pages 191--218, 2019.
+
+\end{thebibliography}
