+++ docs/manual/manual.out
@@ -1,0 +1,27 @@
+\BOOKMARK [1][-]{section.1}{Introduction}{}% 1
+\BOOKMARK [1][-]{section.2}{crobat Features}{}% 2
+\BOOKMARK [2][-]{subsection.2.1}{Orderbook updating}{section.2}% 3
+\BOOKMARK [1][-]{section.3}{Getting Started and Installation}{}% 4
+\BOOKMARK [2][-]{subsection.3.1}{CoPrA}{section.3}% 5
+\BOOKMARK [2][-]{subsection.3.2}{Installation}{section.3}% 6
+\BOOKMARK [1][-]{section.4}{Time-series Data Structure Layout}{}% 7
+\BOOKMARK [2][-]{subsection.4.1}{Introduction}{section.4}% 8
+\BOOKMARK [2][-]{subsection.4.2}{Understanding The Raw Order Book Data}{section.4}% 9
+\BOOKMARK [3][-]{subsubsection.4.2.1}{Order Book Snapshots}{subsection.4.2}% 10
+\BOOKMARK [3][-]{subsubsection.4.2.2}{Signed Order Book}{subsection.4.2}% 11
+\BOOKMARK [2][-]{subsection.4.3}{Final notes on Orderbook interpretation}{section.4}% 12
+\BOOKMARK [1][-]{section.5}{Usage}{}% 13
+\BOOKMARK [1][-]{section.6}{Modules}{}% 14
+\BOOKMARK [2][-]{subsection.6.1}{recorder\137full}{section.6}% 15
+\BOOKMARK [3][-]{subsubsection.6.1.1}{Description}{subsection.6.1}% 16
+\BOOKMARK [3][-]{subsubsection.6.1.2}{class L2\137Update }{subsection.6.1}% 17
+\BOOKMARK [2][-]{subsection.6.2}{LOB\137funcs}{section.6}% 18
+\BOOKMARK [3][-]{subsubsection.6.2.1}{Description:}{subsection.6.2}% 19
+\BOOKMARK [3][-]{subsubsection.6.2.2}{class history\(\)}{subsection.6.2}% 20
+\BOOKMARK [2][-]{subsection.6.3}{history\137funcs.py}{section.6}% 21
+\BOOKMARK [1][-]{section.7}{Credits}{}% 22
+\BOOKMARK [2][-]{subsection.7.1}{Development Lead}{section.7}% 23
+\BOOKMARK [2][-]{subsection.7.2}{Contributors}{section.7}% 24
+\BOOKMARK [1][-]{section.8}{License}{}% 25
+\BOOKMARK [1][-]{section.9}{Appendix}{}% 26
+\BOOKMARK [1][-]{section.10}{To do list}{}% 27
