+++ docs/manual/manual.toc
@@ -1,0 +1,57 @@
+\contentsline {section}{\numberline {1}Introduction}{2}{section.1}%
+\contentsline {section}{\numberline {2}crobat Features}{3}{section.2}%
+\contentsline {subsection}{\numberline {2.1}Orderbook updating}{3}{subsection.2.1}%
+\contentsline {section}{\numberline {3}Getting Started and Installation}{4}{section.3}%
+\contentsline {subsection}{\numberline {3.1}CoPrA}{4}{subsection.3.1}%
+\contentsline {subsection}{\numberline {3.2}Installation}{4}{subsection.3.2}%
+\contentsline {section}{\numberline {4}Time-series Data Structure Layout}{5}{section.4}%
+\contentsline {subsection}{\numberline {4.1}Introduction}{5}{subsection.4.1}%
+\contentsline {subsection}{\numberline {4.2}Understanding The Raw Order Book Data}{5}{subsection.4.2}%
+\contentsline {subsubsection}{\numberline {4.2.1}Order Book Snapshots}{5}{subsubsection.4.2.1}%
+\contentsline {subsubsection}{\numberline {4.2.2}Signed Order Book}{6}{subsubsection.4.2.2}%
+\contentsline {paragraph}{\numberline {4.2.2.1}Signed Order Book Snapshot Prices}{6}{paragraph.4.2.2.1}%
+\contentsline {paragraph}{\numberline {4.2.2.2}Signed Events}{6}{paragraph.4.2.2.2}%
+\contentsline {subsection}{\numberline {4.3}Final notes on Orderbook interpretation}{8}{subsection.4.3}%
+\contentsline {section}{\numberline {5}Usage}{9}{section.5}%
+\contentsline {section}{\numberline {6}Modules}{11}{section.6}%
+\contentsline {subsection}{\numberline {6.1}recorder\_full}{11}{subsection.6.1}%
+\contentsline {subsubsection}{\numberline {6.1.1}Description}{11}{subsubsection.6.1.1}%
+\contentsline {subsubsection}{\numberline {6.1.2}\textit {class} \leavevmode {\color {olive}\texttt {L2\_Update}} }{11}{subsubsection.6.1.2}%
+\contentsline {paragraph}{\numberline {6.1.2.1}\textit {function} \leavevmode {\color {blue}\texttt {\_\_init\_\_}}\texttt {(self, loop, channel, input\_args)}}{12}{paragraph.6.1.2.1}%
+\contentsline {paragraph}{\numberline {6.1.2.2}\textit {function} \leavevmode {\color {blue}\texttt {on\_open}}\texttt {(self)}}{13}{paragraph.6.1.2.2}%
+\contentsline {paragraph}{\numberline {6.1.2.3}\textit {function} \leavevmode {\color {blue}\texttt {on\_message}}\texttt {(self, msg)},}{14}{paragraph.6.1.2.3}%
+\contentsline {paragraph}{\numberline {6.1.2.4}\textit {function} \leavevmode {\color {blue}\texttt {on\_close}}\texttt {(self, was\_clean, code, reason)}}{15}{paragraph.6.1.2.4}%
+\contentsline {subsection}{\numberline {6.2}LOB\_funcs}{16}{subsection.6.2}%
+\contentsline {subsubsection}{\numberline {6.2.1}\textbf {Description:}}{16}{subsubsection.6.2.1}%
+\contentsline {subsubsection}{\numberline {6.2.2}\textit {class} \leavevmode {\color {olive}\texttt {history()}}}{16}{subsubsection.6.2.2}%
+\contentsline {paragraph}{\numberline {6.2.2.1}\textit {function} \leavevmode {\color {blue}\texttt {\_\_init\_\_}}\texttt {(self)}}{16}{paragraph.6.2.2.1}%
+\contentsline {paragraph}{\numberline {6.2.2.2}\textit {function} \leavevmode {\color {blue}\texttt {initialize\_snap\_events}}\texttt {(self, msg, time)}}{17}{paragraph.6.2.2.2}%
+\contentsline {paragraph}{\numberline {6.2.2.3}add\_market\_order\_message}{18}{paragraph.6.2.2.3}%
+\contentsline {paragraph}{\numberline {6.2.2.4}\textit {function} \leavevmode {\color {blue}\texttt {remove\_price\_level}}\texttt {(snap\_array, level\_depth, match\_index)}}{20}{paragraph.6.2.2.4}%
+\contentsline {paragraph}{\numberline {6.2.2.5}\textit {function} \leavevmode {\color {blue}\texttt {update\_level\_depth}}\texttt {(snap\_array, level\_depth, match\_index, pre\_level\_depth)}}{21}{paragraph.6.2.2.5}%
+\contentsline {paragraph}{\numberline {6.2.2.6}\textit {function} \leavevmode {\color {blue}\texttt {update\_price\_index\_buy}}\texttt {(self, level\_depth, price\_level, pre\_level\_depth)}}{22}{paragraph.6.2.2.6}%
+\contentsline {paragraph}{\numberline {6.2.2.7}\textit {function} \leavevmode {\color {blue}\texttt {update\_price\_index\_sell}}\texttt {(self, level\_depth, price\_level, pre\_level\_depth)}}{23}{paragraph.6.2.2.7}%
+\contentsline {paragraph}{\numberline {6.2.2.8}\textit {function} \leavevmode {\color {blue}\texttt {update\_snapshot\_bid}}\texttt {(self)}}{24}{paragraph.6.2.2.8}%
+\contentsline {paragraph}{\numberline {6.2.2.9}\textit {function} \leavevmode {\color {blue}\texttt {update\_snapshot\_ask}}\texttt {(self)}}{25}{paragraph.6.2.2.9}%
+\contentsline {paragraph}{\numberline {6.2.2.10}\textit {function} \leavevmode {\color {blue}\texttt {trim\_coordinator}}\texttt {(self, position, bound)}}{26}{paragraph.6.2.2.10}%
+\contentsline {paragraph}{\numberline {6.2.2.11}\textit {function} \leavevmode {\color {blue}\texttt {append\_snapshot\_bid}}\texttt {(self, time, price\_level)}}{27}{paragraph.6.2.2.11}%
+\contentsline {paragraph}{\numberline {6.2.2.12}\textit {function} \leavevmode {\color {blue}\texttt {append\_snapshot\_ask}}\texttt {(self, time, price\_level)}}{28}{paragraph.6.2.2.12}%
+\contentsline {paragraph}{\numberline {6.2.2.13}\textit {function} \leavevmode {\color {blue}\texttt {append\_signed\_book}}\texttt {(self, time, price\_level)}}{29}{paragraph.6.2.2.13}%
+\contentsline {paragraph}{\numberline {6.2.2.14}\textit {function} \leavevmode {\color {blue}\texttt {chk\_mkt\_can\_overlap}}\texttt {(self, events, order\_type)}}{30}{paragraph.6.2.2.14}%
+\contentsline {paragraph}{\numberline {6.2.2.15}\textit {function} \leavevmode {\color {blue}\texttt {last\_inserted\_order}}\texttt {(self, side=``signed'')}}{31}{paragraph.6.2.2.15}%
+\contentsline {subsection}{\numberline {6.3}history\_funcs.py}{32}{subsection.6.3}%
+\contentsline {paragraph}{\numberline {6.3.0.1}\textit {function} \leavevmode {\color {blue}\texttt {check\_order}}\texttt {(snapshot, side)}}{32}{paragraph.6.3.0.1}%
+\contentsline {paragraph}{\numberline {6.3.0.2}\textit {function} \leavevmode {\color {blue}\texttt {conver\_array\_to\_list\_dict}}\texttt {(history, pos\_limt=5)}}{33}{paragraph.6.3.0.2}%
+\contentsline {paragraph}{\numberline {6.3.0.3}\textit {function} \leavevmode {\color {blue}\texttt {conver\_array\_to\_list\_dict\_sob}}\texttt {(history, events, pos\_limt=5)}}{34}{paragraph.6.3.0.3}%
+\contentsline {paragraph}{\numberline {6.3.0.4}\textit {function} \leavevmode {\color {blue}\texttt {pd\_excel\_save}}\texttt {(title, hist\_obj\_dict)}}{35}{paragraph.6.3.0.4}%
+\contentsline {paragraph}{\numberline {6.3.0.5}\textit {function} \leavevmode {\color {blue}\texttt {set\_sign}}\texttt {(event\_size, side, order\_type)}}{36}{paragraph.6.3.0.5}%
+\contentsline {paragraph}{\numberline {6.3.0.6}\textit {function} \leavevmode {\color {blue}\texttt {set\_signed\_positiion}}\texttt {(position, side)}}{37}{paragraph.6.3.0.6}%
+\contentsline {paragraph}{\numberline {6.3.0.7}\textit {function} \leavevmode {\color {blue}\texttt {get\_min\_dec}}\texttt {(min\_currency\_denom, min\_asset\_value)}}{38}{paragraph.6.3.0.7}%
+\contentsline {section}{\numberline {7}Credits}{39}{section.7}%
+\contentsline {subsection}{\numberline {7.1}Development Lead}{39}{subsection.7.1}%
+\contentsline {subsection}{\numberline {7.2}Contributors}{39}{subsection.7.2}%
+\contentsline {section}{\numberline {8}License}{40}{section.8}%
+\contentsline {section}{\numberline {9}Appendix}{41}{section.9}%
+\contentsline {section}{\numberline {10}To do list}{41}{section.10}%
+\contentsline {paragraph}{\numberline {10.0.0.1}\textit {function} \leavevmode {\color {blue}\texttt {\_\_my\_function\_\_}}\texttt {(self, param\_1, param\_2=False, param\_3)}}{41}{paragraph.10.0.0.1}%
+\contentsline {paragraph}{\numberline {10.0.0.2}\textit {function} \leavevmode {\color {blue}\texttt {\_\_my\_function\_\_}}\texttt {(self, param\_1, param\_2=False, param\_3)}}{42}{paragraph.10.0.0.2}%
